--- command-core/src/main/kotlin/com/cereal/script/data/json/Json.kt
@@ -1,9 +1,0 @@
-package com.cereal.script.data.json
-
-import kotlinx.serialization.json.Json
-
-fun defaultJson() =
-    Json {
-        ignoreUnknownKeys = true
-        isLenient = true
-    }
