+++ command-monitoring/src/main/kotlin/com/cereal/command/monitor/data/common/json/Json.kt
@@ -1,0 +1,9 @@
+package com.cereal.command.monitor.data.common.json
+
+import kotlinx.serialization.json.Json
+
+fun defaultJson() =
+    Json {
+        ignoreUnknownKeys = true
+        isLenient = true
+    }
