+++ command-monitoring/src/main/kotlin/com/cereal/command/monitor/data/shopify/ShopifyItemRepository.kt
@@ -1,14 +1,14 @@
 package com.cereal.command.monitor.data.shopify
 
+import com.cereal.command.monitor.data.common.httpclient.defaultHttpClient
+import com.cereal.command.monitor.data.common.json.defaultJson
+import com.cereal.command.monitor.data.common.useragent.MOBILE_USER_AGENTS
 import com.cereal.command.monitor.data.shopify.models.ShopifyResponse
 import com.cereal.command.monitor.models.Item
 import com.cereal.command.monitor.models.ItemProperty
 import com.cereal.command.monitor.models.Page
 import com.cereal.command.monitor.models.Variant
 import com.cereal.command.monitor.repository.ItemRepository
-import com.cereal.script.data.httpclient.defaultHttpClient
-import com.cereal.script.data.json.defaultJson
-import com.cereal.script.data.useragent.MOBILE_USER_AGENTS
 import com.cereal.script.repository.LogRepository
 import com.cereal.sdk.models.proxy.RandomProxy
 import io.ktor.client.request.get
