+++ command/build.gradle.kts
@@ -17,6 +17,7 @@ dependencies {
     testImplementation("org.junit.jupiter:junit-jupiter:5.10.0")
     testImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.9.0")
     testImplementation("io.mockk:mockk:1.13.13")
+    testImplementation("com.cereal-automation:cereal-test-utils:1.4.0")
 
     // Ktor
     implementation("io.ktor:ktor-client-core:2.3.13")
@@ -27,8 +28,8 @@ dependencies {
     implementation("io.ktor:ktor-client-auth:2.3.13")
     implementation("io.ktor:ktor-client-encoding:2.3.13")
 
-    implementation(project(":command"))
-    testImplementation(project(":command"))
+    implementation("org.jetbrains.kotlinx:kotlinx-datetime:0.6.1")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.9.0")
 }
 
 tasks.test {
