+++ command/src/main/kotlin/com/cereal/script/commands/Command.kt
