+++ command/src/test/kotlin/com/cereal/script/fixtures/FooCommand.kt
