+++ script-bdga-store/src/main/kotlin/com/cereal/bdgastore/BDGAStoreScript.kt
@@ -4,10 +4,10 @@ import com.cereal.command.monitor.MonitorCommandFactory
 import com.cereal.command.monitor.data.shopify.ShopifyItemRepository
 import com.cereal.command.monitor.data.shopify.ShopifyWebsite
 import com.cereal.command.monitor.strategy.StockAvailableMonitorStrategy
+import com.cereal.script.CommandExecutionScript
 import com.cereal.sdk.ExecutionResult
 import com.cereal.sdk.Script
 import com.cereal.sdk.component.ComponentProvider
-import com.cereal.shared.CommandExecutionScript
 import kotlin.time.Duration.Companion.seconds
 
 class BDGAStoreScript : Script<BDGAStoreConfiguration> {
