+++ script-common/src/main/resources/manifest.json
@@ -1,0 +1,5 @@
+{
+  "package_name": "com.cereal-automation.monitor.nike",
+  "name": "Nike Product Monitor",
+  "version_code": 1
+}
