+++ script-nike/src/main/kotlin/com/cereal/nike/NikeConfiguration.kt
@@ -1,9 +1,9 @@
 package com.cereal.nike
 
 import com.cereal.command.monitor.data.nike.ScrapeCategory
+import com.cereal.script.utils.configuration.BaseConfiguration
 import com.cereal.sdk.ScriptConfigurationItem
 import com.cereal.sdk.models.proxy.RandomProxy
-import com.cereal.shared.BaseConfiguration
 
 interface NikeConfiguration : BaseConfiguration {
     @ScriptConfigurationItem(
