+++ script-nike/src/main/kotlin/com/cereal/nike/NikeScript.kt
@@ -5,11 +5,11 @@ import com.cereal.command.monitor.data.nike.NikeItemRepository
 import com.cereal.command.monitor.strategy.MonitorStrategy
 import com.cereal.command.monitor.strategy.NewItemAvailableMonitorStrategy
 import com.cereal.command.monitor.strategy.PriceDropMonitorStrategy
+import com.cereal.script.CommandExecutionScript
 import com.cereal.script.commands.Command
 import com.cereal.sdk.ExecutionResult
 import com.cereal.sdk.Script
 import com.cereal.sdk.component.ComponentProvider
-import com.cereal.shared.CommandExecutionScript
 import kotlinx.datetime.Clock
 import kotlin.time.Duration.Companion.seconds
