--- script-shared/build.gradle.kts
@@ -1,23 +1,0 @@
-plugins {
-    kotlin("jvm")
-}
-
-dependencies {
-    api(project(":command-core"))
-    api(project(":command-monitoring"))
-    implementation("org.jetbrains.kotlinx:kotlinx-datetime:0.6.1")
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.9.0")
-
-    testImplementation(kotlin("test"))
-    testImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.9.0")
-    testImplementation("io.mockk:mockk:1.13.13")
-    testImplementation("com.cereal-automation:cereal-test-utils:1.4.0")
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
-
-kotlin {
-    jvmToolchain(17)
-}
