--- script-shared/src/test/kotlin/com/cereal/shared/InMemoryScriptConfig.kt
@@ -1,10 +1,0 @@
-package com.cereal.shared
-
-import com.cereal.sdk.statemodifier.ScriptConfig
-import com.cereal.sdk.statemodifier.ScriptConfigValue
-
-class InMemoryScriptConfig(
-    private val configuration: Map<String, ScriptConfigValue>,
-) : ScriptConfig {
-    override fun valueForKey(key: String): ScriptConfigValue = configuration[key] ?: ScriptConfigValue.NullScriptConfigValue
-}
