--- script-shared/src/test/kotlin/com/cereal/shared/fixtures/FakeNotificationRepository.kt
@@ -1,13 +1,0 @@
-package com.cereal.shared.fixtures
-
-import com.cereal.command.monitor.models.Item
-import com.cereal.command.monitor.repository.NotificationRepository
-
-class FakeNotificationRepository : NotificationRepository {
-    override suspend fun notify(
-        message: String,
-        item: Item,
-    ) {
-        // No-op
-    }
-}
