+++ script-snkrs/src/main/kotlin/com/cereal/snkrs/SnkrsConfiguration.kt
@@ -1,9 +1,9 @@
 package com.cereal.snkrs
 
 import com.cereal.command.monitor.data.snkrs.Locale
+import com.cereal.script.utils.configuration.BaseConfiguration
 import com.cereal.sdk.ScriptConfigurationItem
 import com.cereal.sdk.models.proxy.RandomProxy
-import com.cereal.shared.BaseConfiguration
 
 interface SnkrsConfiguration : BaseConfiguration {
     @ScriptConfigurationItem(
