+++ settings.gradle.kts
@@ -2,10 +2,10 @@ plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.5.0"
 }
 rootProject.name = "Script Monitoring"
-include("command-core")
-include("script-shared")
+include("command")
+include("command-monitoring")
+include("script-common")
 include("script-sample")
 include("script-nike")
 include("script-snkrs")
 include("script-bdga-store")
-include("command-monitoring")
